Only close classloaders on reload/fail
We should only close the classloader during reload, disable should not
close the classloader of the plugin other than during a reload or when
a plugin fails to load due to plugins (mis)using this method to reload
their plugins.

This behavior is also exposed by the API, allowing for plugins to disable
other plugins and control if their classloader should also be closed.