Configurable option to increase tab spam limiter
In 1.13, clients will send a tab competion packet to the server each
character a player types for a command with an unknown completion,
This can cause varying issues for commands such as messaging.

While this patch may allow you to increase, or even disable the limit,
doing so will leave the door further open for people to potentially cause
performance issues on servers, especially when used with some plugins.